fix: creating qcow2 on x86_64

+++ create_qcow2.sh
@@ -21,11 +21,16 @@ then
 fi
 
 cd ${LIMA_HOME}/${VM_NAME}
-qemu-img convert -f raw -O qcow2 -o backing_file=basedisk,backing_fmt=qcow2 \
-    diffdisk diffdisk-linked.qcow2
-qemu-img convert -p -O qcow2 diffdisk-linked.qcow2 pwnpad-${ARCH}.qcow2
 
-rm diffdisk-linked.qcow2
+if [ ${ARCH} == "aarch64" || ${ARCH} == "arm64" ]; then
+    qemu-img convert -f raw -O qcow2 -o backing_file=basedisk,backing_fmt=qcow2 \
+        diffdisk diffdisk-linked.qcow2
+    qemu-img convert -p -O qcow2 diffdisk-linked.qcow2 pwnpad-${ARCH}.qcow2
+
+    rm diffdisk-linked.qcow2
+else if [ ${ARCH} == "x86_64" ]; then
+    qemu-img convert -p -O qcow2 diffdisk pwnpad-${ARCH}.qcow2
+fi
 
 virt-sparsify --compress pwnpad-${ARCH}.qcow2 pwnpad-new.qcow2
 mv pwnpad-new.qcow2 pwnpad-${ARCH}.qcow2
